Add script to discover inconsistencies between the Plone Lexicon and ZCTextIndexes.
This script attempts to discover inconsistencies between the Plone Lexicon
and ZCTextIndexes.

It should detect cases where a ZCTextIndex references a word (by an integer
word ID (wid)) that doesn't exist in the lexicon (any more). It does this
by trying to fetch all indexed data, for all indexes, for all brains.

In the case of ZCTextIndexes, this will cause `getEntryForObject` to look up
a list of word IDs in the lexicon, and trigger a KeyError if that word isn't
in the lexicon.

It seems that the only reliable way to fix this situation is to first CLEAR
all the affected indexes, and then rebuild them.