Script for checking datetime indexes for falsy values.
A bin/instance run script which reports catalog datetime indexes
which contain falsy values.

A datetime index must not contain falsy values, otherwise the affected
brains will be removed from a result set when ordered by this index.

Background:
Sorting by a DateTime index containing falsy values is problematic:
the index has not enough infos to decide whether the falsy entries should
be ordered on top or on bottom of the list.
The index simply removes the falsy values.

ftw.table's catalog source contains a workaround which prevents the catalog
from sorting by date indexes in general and does this by hand afterwards.
This is a bad solution.

The correct solution is to extend the indexers of date indexes so that they
always return a datetime. The indexer must decide whether this is a very-early
or very-late datetime object.

With this script we proof that our indexes are correct, so that we can fix this
by removing ftw.table workaround in GEVER.