Add fix_one_note_mimetype script.
With https://github.com/4teamwork/opengever.core/pull/3660 the
mimetype for onenote document has been fixe. But the fix in 3660
solves the problem only for new but not for existing
documents. Because we don't have an index for the contentType or the
filename an upgradestep would be way to longrunning comparing how many
documents are effected. Therefore we add this simple
opengever.maintenance script. We can run offside the release.