Add script to list decided proposals without excerpt.
We add a script that lists the decided proposals in
closed meetings, but lacking a returned excerpt. As
it is now mandatory to return an excerpt for each proposal
before closing a meeting, this allows us to get an idea of
how many proposals are not conforming to this new rule.