Avoid formating error with empty tuples.
The format modulo operator has a special handling for tuples as they can be used
to pass multiple arguments to the operator, see
https://docs.python.org/2/library/stdtypes.html#string-formatting-operations. We
don't want this for default and missing values as they can be an empty typle, so
we need to wrap tuples in an extra tuple.

I can't resist to say this: with the format method this wouldn't happen :-p.