Handle ambiguous dossier / task states not listed in their respective
OPEN or CLOSED STATES mappings:

For tasks in states skipped or planned we consider them closed for the purpose
of this script. They can't be edited or added to, and planned tasks can't
exist without being in an open task anyway, which is already prevented.

The "offered" state for dossiers is not handled explicitly for now. It probably
would also be considered closed, but this would have to be looked at in
combination with disposition objects (which are required for a dossier to
even be offered), and the existence of those would already block the script
because of their unexpected portal_type.

Therefore we don't address the whole disposition aspects yet, because that's
not a use case yet for the customer where this script will be executed.

A check is added however that every encountered review state is mapped as
either open or closed, so that unexpected states don't go unnoticed in the
future.