Repo pos can contain dossier while regenerating refnum mapping.
It is possible that all child repository folders get moved out
of a repository folder during a migration and that other
repository folders get merged into it, so that it was a parent
of a repository folder before the migration but now contains
dossiers. In that case the mapping still needs to be regenerated
but we cannot throw an error if it contains dossiers.