Configure InterGEVER webactions according to new permission model:
The InterGEVER webactions are now protected with the respective groups
according to the new InterGEVER permission model.

This requires that the deployment definitions in this script contain an
"administrators_group" key, as well as a "connected_admin_units" that contains
the definitions of the connected admin units and their groups.